Fixed Milestone 2 Errors
I was mixing up % and / for the row and collum. I fixed that for all of
the places that used % and / as row and collom. This fixed the error
that the powerup was sometimes under the player. Also note it is not a
bug if a powerup goes missing, it is just under a ninja and the Ninja is
more important to show than the power up. Also fixed the part where
ninjas could not show near 2 spaces of player. Originally they were made
they could not appear 3 from left and 3 from bottom. I changed that to 3
from left or 3 from bottom so now the ninja can appear touching the
bottom wall as long as he is 3 away from the left (making him the needed
two spaces away from the player).