Fully Commented Code without saving or loading
All that needs to be done is saving and loading. Empty functions
saveGame() and loadGame() are already in the UI. I am done coding for
this project unless you guys really need my help. (I feel as if I have
done more than my fair share of the work) If there is a bug with this
code please tell me and ill see if can fix it. Good Luck.